refactor: remove deprecated loading and skeleton components, replace with unified ListSkeleton
