+++ apps/web/app/(app)/top/artists/page.tsx
@@ -2,12 +2,12 @@ import { Suspense } from "react";
 
 import { getTopArtistsAction } from "~/actions/get-top-user-action";
 import { AppHeader } from "~/components/app-header";
+import { ListSkeleton } from "~/components/list-skeleton";
 import { SelectTimeRange } from "~/components/select-time-range";
 import { SelectTopLayout } from "~/components/select-top-layout";
 import { getCookieTopTimeRange } from "~/lib/utils-server";
 
 import { TopArtistList } from "../components/top-artists-list";
-import { TopListSkeleton } from "../components/top-list-skeleton";
 
 export default function TopArtistsPage() {
   return (
@@ -17,7 +17,7 @@ export default function TopArtistsPage() {
         <SelectTopLayout />
       </AppHeader>
       <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
-        <Suspense fallback={<TopListSkeleton layout="list" />}>
+        <Suspense fallback={<ListSkeleton />}>
           <TopArtistListWrapper />
         </Suspense>
       </div>
