+++ apps/web/app/(app)/top/tracks/page.tsx
@@ -2,11 +2,11 @@ import { Suspense } from "react";
 
 import { getTopTracksAction } from "~/actions/get-top-user-action";
 import { AppHeader } from "~/components/app-header";
+import { ListSkeleton } from "~/components/list-skeleton";
 import { SelectTimeRange } from "~/components/select-time-range";
 import { SelectTopLayout } from "~/components/select-top-layout";
 import { getCookieTopTimeRange } from "~/lib/utils-server";
 
-import { TopListSkeleton } from "../components/top-list-skeleton";
 import { TopTracksList } from "../components/top-tracks-list";
 export default async function TopTracksPage() {
   return (
@@ -16,7 +16,7 @@ export default async function TopTracksPage() {
         <SelectTopLayout />
       </AppHeader>
       <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
-        <Suspense fallback={<TopListSkeleton layout="list" />}>
+        <Suspense fallback={<ListSkeleton />}>
           <TopTracksListWrapper />
         </Suspense>
       </div>
